// src/interfaces/Enum.ts
export enum TypeFile {
  HR = "hr",
  Training = "training",
  Equipment = "equipment",
  Monitory = "monitory",
}
export enum TrapType {
  mouseTrap = "mouseTrap",
  insectTrap = "insectTrap",
}
export enum Role {
  Manager = "manager",
  Censor = "censor",
  Staff = "staff",
}

export enum StatusFile {
  pending = "pending",
  reject = "reject",
  passCensor = "passCensor",
  passManager = "passManager",
}

export enum Gender {
  Male = "male",
  Female = "female",
}